# docs/swagger.yaml
definitions:
  age.AgeCategory:
    properties:
      id:
        type: integer
      max_age:
        type: integer
      min_age:
        type: integer
      name:
        type: string
      note:
        type: string
      user_id:
        type: integer
    required:
    - max_age
    - min_age
    - name
    type: object
...
  routes.AgeCategoryRequest:
    properties:
      max_age:
        type: integer
      min_age:
        type: integer
      name:
        type: string
      note:
        type: string
    required:
    - max_age
    - min_age
    - name
    type: object
...
  routes.ChangePasswordRequest:
    properties:
      confirmPassword:
        type: string
      newPassword:
        type: string
      oldPassword:
        type: string
    required:
    - confirmPassword
    - newPassword
    - oldPassword
    type: object
  routes.DefaultMessageResponse:
    properties:
      message:
        type: string
    type: object
...
  routes.SendCodeRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  routes.VerifyCodeRequest:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
info:
  contact: {}
paths:
  /api/age-categories:
    get:
      description: Retrieves all age categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/age.AgeCategory'
            type: array
        "404":
          description: Age categories not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      summary: Get all age categories
    post:
      description: Creates a new age category
      parameters:
      - description: Age Category
        in: body
        name: ageCategory
        required: true
        schema:
          $ref: '#/definitions/routes.AgeCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Age category created
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "409":
          description: Age category already exists
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new age category
  /api/age-categories/{id}:
    delete:
      description: Deletes an age category with the specified ID
      parameters:
      - description: Age Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category Deleted
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Age category Not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "409":
          description: Age category is Used in movies
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete an age category
    get:
      description: Retrieves an age category with the specified ID
      parameters:
      - description: Age Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/age.AgeCategory'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Age category not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      summary: Get an age category by ID
    put:
      description: Updates an age category with the specified ID
      parameters:
      - description: Age Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Age Category
        in: body
        name: ageCategory
        required: true
        schema:
          $ref: '#/definitions/routes.AgeCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category Updated
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Age category Not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "409":
          description: Age Category with this name already exists
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Update an age category
...
  /api/change-password:
    post:
      consumes:
      - application/json
      description: Changes the password of the authenticated user
      parameters:
      - description: Password change data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/routes.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Change user password
      tags:
      - passwordChange
...
  /api/send-code:
    post:
      consumes:
      - application/json
      description: Sends a verification code to the provided email address
      parameters:
      - description: Email data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/routes.SendCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Code sent
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      summary: Send verification code
      tags:
      - passwordChange
...
  /api/verify-code:
    post:
      consumes:
      - application/json
      description: Verifies the verification code for the provided email address and
        generates a temporary password
      parameters:
      - description: Verification code data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/routes.VerifyCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Temporary password sent
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      summary: Verify verification code
      tags:
      - passwordChange
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"